+++ src/execution/use-timers.ts
@@ -1,12 +1,16 @@
 import { timers } from '../globalization';
+import { startTimer } from './start-timer';
+import { stopTimer } from './stop-timer';
 
-export function useTimers(state: 'on' | 'off') {
+type TimerState = 'on' | 'off'
+
+export function useTimers(state: TimerState): void;
+export function useTimers(active: boolean): void;
+export function useTimers(arg: TimerState | boolean) {
+  const state: TimerState = (typeof arg === 'boolean' ? (arg ? 'on' : 'off') : arg );
   if (state === 'on') {
-    timers.forEach(timer => {
-      const { method, interval, target } = timer;
-      timer.handle = setInterval(() => target ? method.apply(target) : method(), interval || 60000);
-    });
+    timers.forEach(timer => startTimer(timer.id));
   } else if (state === 'off') {
-    timers.forEach(timer => clearInterval(timer.handle));
+    timers.forEach(timer => stopTimer(timer.id));
   }
 }